#51 - feat : [BE] loading API 개발

+++ backend/src/main/java/org/ssafy/load/dto/response/LoadTaskGoodsResponse.java
@@ -3,9 +3,9 @@ package org.ssafy.load.dto.response;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record LoadTaskGoodsResponse(
-        @JsonProperty("goods_id") int goodsId,
+        @JsonProperty("goods_id") long goodsId,
         String type,
         int weight,
-        @JsonProperty("building_id") int buildingId
+        @JsonProperty("building_id") long buildingId
 ) {
 }
